modules/API-IMDB/requirements.txt
flask
flask_cors
requests 
uuid
mongoengine
coloredlogs
dnspython
flask-restplus
pymongo
imdbpy
